fix: Move piece when checking if the move is valid
Before the fix, when checking if the move is valid with #will_in_check?,
it will clone the game object and then move the piece in the cloned
object to verify if the cloned object will lead to player's king in
check.

However, moving piece in the cloned object will also affect the original
object because the #clone method only do the shallow copy.

In the fix, updated the #will_in_check? to check whether a move will
make the player's king in check by simulate the move, save whether it is
in check in the result variable, revert the move, and show the result.