fix: en passant is treated as invalid move
Before the fix, en passant cannot be performed because it is treated as
invalid move. When checking whether the en passant is allowed, the
program will check if the enemy makes a double step before with
@advanced_2_squares variable. However, the @advanced_2_squares was not
updated before the fix. It leads the allow_en_passant? always return
false and therefore the en passant cannot be performed.

In the fix, @advanced_2_squares will be updated to true if the pawn's
last move is a double step. Otherwise, the @advanced_2_squares of the
pawn will remain false.