fix: Cannot check if the move make player in check
Before the fix, error is not shown when player makes an illegal move
that makes himself in check. To check if the move makes player in check,
the program will check if any enemy piece can capture the player's king
in next step according to the king's current position.

However, in #will_in_check?, piece's current position is not updated.
When moving the king to an position that will make player in check, the
program is checking the original position of the king instead of the new
position and it leads to the incorrect #will_in_check? result.

Thus, the current position of the piece will be temporarily updated for
checking if the new position makes the king in check. And will revert
the position afterward.